refactor: create new datasource module
Create a new Datasource module that will host the various data sources. For now move the activity
into there instead of server, in the future more will be moved across to this so that the api and
the local data sources can be combined into this module
