+++ app/build.gradle
@@ -67,6 +67,8 @@ dependencies {
     implementation project(":libraries:datasource:series")
     implementation project(":libraries:datasource:trending")
     implementation project(":libraries:datasource:user")
+    implementation project(":libraries:datasource:activity")
+    implementation project(":libraries:encryption")
     implementation project(":libraries:kitsu")
     implementation project(":libraries:kitsu:activity")
     implementation project(":libraries:kitsu:auth")
@@ -75,7 +77,6 @@ dependencies {
     implementation project(":libraries:kitsu:trending")
     implementation project(":libraries:kitsu:user")
     implementation project(":libraries:library")
-    implementation project(":libraries:server:activity")
     implementation project(":libraries:server:auth")
     implementation project(":libraries:server:library")
     implementation project(":libraries:server:search")
