+++ libraries/datasource/activity/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.chesire.nekome.datasource.activity" />
