+++ libraries/datasource/activity/src/main/java/com/chesire/nekome/datasource/activity/ActivityDomain.kt
@@ -1,4 +1,4 @@
-package com.chesire.nekome.activity.api
+package com.chesire.nekome.datasource.activity
 
 /**
  * Provides a domain for a single piece of activity that has occurred.
