+++ libraries/datasource/activity/src/main/java/com/chesire/nekome/datasource/activity/ActivityRepository.kt
@@ -1,0 +1,4 @@
+package com.chesire.nekome.datasource.activity
+
+class ActivityRepository {
+}
