+++ libraries/datasource/activity/src/main/java/com/chesire/nekome/datasource/activity/remote/ActivityApi.kt
@@ -1,6 +1,7 @@
-package com.chesire.nekome.activity.api
+package com.chesire.nekome.datasource.activity.remote
 
 import com.chesire.nekome.core.Resource
+import com.chesire.nekome.datasource.activity.ActivityDomain
 
 /**
  * Methods relating to retrieving the users activity.
