+++ libraries/kitsu/activity/build.gradle
@@ -31,8 +31,8 @@ android {
 
 dependencies {
     implementation project(":libraries:core")
+    implementation project(':libraries:datasource:activity')
     implementation project(':libraries:kitsu')
-    implementation project(':libraries:server:activity')
 
     implementation "com.chesire.lintrules:lint-gradle:$lintrules_version"
     implementation "com.chesire.lintrules:lint-xml:$lintrules_version"
