+++ libraries/kitsu/activity/src/main/java/com/chesire/nekome/kitsu/activity/KitsuActivity.kt
@@ -1,7 +1,7 @@
 package com.chesire.nekome.kitsu.activity
 
-import com.chesire.nekome.activity.api.ActivityApi
-import com.chesire.nekome.activity.api.ActivityDomain
+import com.chesire.nekome.datasource.activity.remote.ActivityApi
+import com.chesire.nekome.datasource.activity.ActivityDomain
 import com.chesire.nekome.core.Resource
 import com.chesire.nekome.kitsu.activity.dto.RetrieveActivityDto
 import com.chesire.nekome.kitsu.asError
