+++ settings.gradle
@@ -14,6 +14,9 @@ include ':app',
     ':libraries:datasource:series',
     ':libraries:datasource:trending',
     ':libraries:datasource:user',
+    ':libraries:datasource',
+    ':libraries:datasource:activity',
+    ':libraries:encryption',
     ':libraries:kitsu',
     ':libraries:kitsu:activity',
     ':libraries:kitsu:auth',
@@ -22,7 +25,6 @@ include ':app',
     ':libraries:kitsu:trending',
     ':libraries:kitsu:user',
     ':libraries:library',
-    ':libraries:server:activity',
     ':libraries:server:auth',
     ':libraries:server:library',
     ':libraries:server:search',
